feat: navigate user to dashboard when they press back on builder page
Currently, when users press the back button on the builder page, if they are not logged in, they are redirected back to the landing page. Users are more likely to close the tab at this point, because they can't see their work. This commit changes the behavior so that users are redirected to the dashboard instead, allowing them to see their saved projects and continue working.