Development (#40)
* Feature: Local Resumes (#34)

* feat: make dashboard page public

* fix: style fixes in dashboard page when not signed in

* feat: allow resume crud operations without signup using IndexedDB

* feat: create instant resume from landing page cta

* fix: resolved build issues

* feat: import resumes without signup, add new features to landing page

* docs: update README with no-signup feature

* change: added new posthog events for anonymous resume creation and import

* feat: sync local resumes with account on login or from documents tab

* feat: navigate user to dashboard when they press back on builder page

Currently, when users press the back button on the builder page, if they are not logged in, they are redirected back to the landing page. Users are more likely to close the tab at this point, because they can't see their work. This commit changes the behavior so that users are redirected to the dashboard instead, allowing them to see their saved projects and continue working.

* refactor: code clean up, change landing page url, add seo changes

* fix: update next js version and made template change instant

* chore: add google site verification meta tag